fix restoreAirLoadouts to only run on SA - move cfgs to sep file
