+++ FNF_MissionTemplate.VR/description/cfgFNFVehicleLoadouts.hpp
@@ -1,0 +1,113 @@
+// automated turret weapon loadout changes
+
+// 2B14
+class rhs_2b14_82mm_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_2B14Podnos";
+  };
+};
+// AH1Z
+class RHS_AH1Z_base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_AH1Z";
+  };
+};
+// Ka52
+class RHS_Ka52_base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_Ka52";
+  };
+};
+  // GAZ Tigr
+class rhs_tigr_base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_GAZ";
+  };
+};
+// M252
+class RHS_M252_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_M252";
+  };
+};
+// AH6M Littlebird
+class RHS_MELB_AH6M {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_AH6M";
+  };
+};
+  // Mi-24P
+class RHS_Mi24P_VVS_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_Mi24P";
+  };
+};
+// Mi-24V, Mi-24Vt
+class RHS_Mi24V_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_Mi24V";
+  };
+};
+// MkVSOC
+class rhsusf_mkvsoc {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_MkVSOC";
+  };
+};
+  // M1117
+class rhsusf_M1117_base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_M1117";
+  };
+};
+// speedboat minigun
+class Boat_Armed_01_minigun_base_F {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_speedboatMinigun";
+  };
+};
+// speedboat HMG
+class O_Boat_Armed_01_hmg_F {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_speedboatHMG";
+  };
+};
+
+// 3CB
+
+// AAV
+class UK3CB_AAV {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_AAV";
+  };
+};
+// BMP-1
+class UK3CB_BMP1Tank_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_BMP1";
+  };
+};
+// BMP-2
+class UK3CB_BMP2Tank_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_BMP2";
+  };
+};
+// LAV-25
+class UK3CB_LAV25 {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_LAV25";
+  };
+};
+// LAV-C2
+class UK3CB_LAV25_HQ {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_LAVC2";
+  };
+};
+// FV-510
+class UK3CB_Warrior_Base {
+  class FNF_SetLoadout {
+    serverInit = "_this call fnf_vehicleLoadouts_fnc_FV510";
+  };
+};
