+++ FNF_MissionTemplate.VR/server/init/fn_restoreAirLoadouts.sqf
@@ -2,9 +2,9 @@ params ["_vehicle"];
 // "debug_console" callExtension "attempt 3265";
 // "debug_console" callExtension format["_this = %1", _this];
 
-// don't run if using FNF default loadout
+// don't run if not SA
+if (missionNamespace getVariable ["fnf_gamemode", "destroy"] != "sustainedAssault") exitWith {};
 waitUntil {!isNull _vehicle};
-if !([_vehicle] call fnf_vehicleLoadouts_fnc_shouldModify) exitWith {};
 
 private _pylonInfo = _vehicle getVariable ["fnf_pylonLoadout", []];
 if (count _pylonInfo == 0) exitWith {};
